refactor(workspace): move the artifact families out of the viewer (§13.1 stage two)
`ArtifactContentViewer.tsx` goes from 6,862 to 3,169 lines. It keeps the
dispatcher, the shared artifact contracts and the entangled raw-attempt and
draft-log sections the section says to leave in place; everything with a clean
boundary now sits in `src/components/workspace/artifactViewers/`.

Shared first, so no slice has to import back into the dispatcher and no cycle
is possible:

- `rawAttempts.ts` — attempt-to-variant construction, dedupe, and log-fallback
  recovery, used by six families.
- `ArtifactProcessingNotice.tsx` + `artifactProcessingMetadata.ts` — the
  notice, and the structured-output merge helpers behind it.
- `RawContentView.tsx`, `CouncilDraftFailureDiagnostics.tsx`.

Then the ten families the section names, plus the integration report, which
sits between two of them and had the same clean boundary:
`coverageViews.tsx` + `coverageSummary.ts`, `relevantFilesView.tsx`,
`executionSetupViews.tsx`, `finalTestView.tsx`, `integrationReportView.tsx`,
`pullRequestViews.tsx`, `cleanupReportView.tsx`, `preFlightReportView.tsx`,
`beadDiffViews.tsx`, `manualQaViews.tsx` + `manualQaArtifact.ts`.

Where a family exported both components and plain functions, the functions
went into a sibling `.ts` module rather than staying alongside the components:
`eslint`'s `react-refresh/only-export-components` is a warning, and AGENTS.md
says warnings get fixed, not tolerated.

Every moved declaration is byte-identical to what it replaced — the moves were
scripted, and the only hand edits are import statements. The dispatcher's
public exports are unchanged, so `PhaseArtifactsPanel`, `ApprovalView`,
`ExecutionSetupPlanApprovalPane` and the artifact tests are untouched.

Impact: no visual, class-name or copy change. 1,244 client tests and 113 test
files pass; `npm run build` succeeds, which is what proves no import cycle was
introduced — a cycle here would have surfaced as a Rollup warning or an
undefined component at module-init time.
