+++ src/components/workspace/artifactViewers/RawContentView.tsx
@@ -1,0 +1,21 @@
+export function RawContentView({ content }: { content: string }) {
+  const lines = content.split('\n')
+  return (
+    <div className="text-sm space-y-1">
+      {lines.map((line, i) => {
+        if (line.startsWith('# ')) return <h2 key={i} className="text-lg font-bold mt-3 mb-1">{line.slice(2)}</h2>
+        if (line.startsWith('## ')) return <h3 key={i} className="text-base font-semibold mt-2 mb-1">{line.slice(3)}</h3>
+        if (line.startsWith('### ')) return <h4 key={i} className="text-sm font-semibold mt-2">{line.slice(4)}</h4>
+        if (line.startsWith('✅')) return <div key={i} className="flex items-center gap-1 text-green-600"><span>{line}</span></div>
+        if (line.startsWith('⚠️')) return <div key={i} className="flex items-center gap-1 text-yellow-600"><span>{line}</span></div>
+        if (line.startsWith('❌')) return <div key={i} className="flex items-center gap-1 text-red-600"><span>{line}</span></div>
+        if (line.startsWith('|')) return <div key={i} className="font-mono text-xs">{line}</div>
+        if (line.startsWith('- ') || line.startsWith('* ')) return <div key={i} className="ml-4">• {line.slice(2)}</div>
+        if (line.match(/^\d+\./)) return <div key={i} className="ml-4">{line}</div>
+        if (line.startsWith('**') && line.endsWith('**')) return <div key={i} className="font-semibold">{line.replace(/\*\*/g, '')}</div>
+        if (line.trim() === '') return <div key={i} className="h-2" />
+        return <div key={i}>{line}</div>
+      })}
+    </div>
+  )
+}
